fixed problem when returning the response from the delete method on products, modified product entity props from 'expiringDate' being a LocalDate type, to 'expirationDate' being a String type

+++ src/main/java/competcommerce/persistence/entity/Client.java
@@ -22,7 +22,7 @@ public class Client {
     private String password;
 
     @Column(nullable = false, name = "phone_number")
-    private int phoneNumber;
+    private Long phoneNumber;
 
     @Column(nullable = false, name = "national_code")
     private short nationalCode;
@@ -31,7 +31,7 @@ public class Client {
             String name,
             String email,
             String password,
-            int phoneNumber,
+            Long phoneNumber,
             short nationalCode
     ) {
         this.name = name;
